requirements.txt
asgiref==3.8.1
Django==5.1.4
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
phonenumbers==8.13.52
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
